Store Student ID in long long instead of int
This fixes an overflow issue when having a 10 digit ID. The reason this
is long long instead of just long is for cross-compatibility. If I
recall correctly, long and int are the same size on Windows. And, since
I haven't switched to using things like int64_t, I'll just set it to
long long.