Added Qt Creator project files
I recently started using Qt Creator, after finding out that it was not
solely for Qt development. Unlike Eclipse or Visual Studio, it does not
create a bunch of project files, only two (when not using CMake). So I
thought I'd include them in the project if anybody else uses it as well.