Debug mode with command line option
Instead of having to recompile the application to enable and disable
outputting debug images, use the -d command line option. This generates
some warnings since DEBUG isn't used in all files options.h is included
in, which needs to be fixed; however, this majorly reduces the run time
if we don't need to be outputting those images.