Fixed command-line ID integer overflow
Earlier I fixed the ID integer overflow internally, but if you passed in
a large teacher ID on the command line, it would overflow. Now
we use 'long long' consistently for the ID everywhere. Also exit when
we would overflow the teacher ID or max thread count, and only init OpenIL
after all this command-line parsing since if we exit we definitely won't
be doing any image processing.