Initial website using CppCMS
This adds a website that supports user account management, form uploads,
and dummy form processing.

I went with CppCMS so that I could manage everything within the freetron
application in different threads. That way after upload I can directly
add the item to a queue to be processed and monitor the result giving a
processing percentage to the user. Not there yet, but this is the first
step.