Show results on website after processing
The build option to not build the website has been removed since now
Processor adds the results to the database in website mode. Also got
proper loading and deleting in the Javascript code.

Note: there is a slight chance that after processing the form when it is
deleted that a website thread may try accessing the form through a
reference. This should be looked into.