Multi-request processing updates with Json RPC
Instead of using a hacked-together percentages followed by line return
and flush, respond with the status update in the Json RPC with a single
update. Start a new request in Javascript if not 100% done to get the
next status update. Not only is this cleaner, but it makes it work
properly in Chrome as well.