Generate keys on install
Now there is no example config, just the real one pointing to hmac.txt
and cbc.txt that will contain the keys. Then on install generate these
keys.

Note: the command to generate the keys uses --hmac-cbc to output to the
cbc file although the cppcms_make_key says it should be --cbc-file. I
reported the bug on Sourceforge.

https://sourceforge.net/p/cppcms/bugs/129/