Moved includes, some messages only for CLI mode
I switched from putting all includes in the header files to putting only
the needed ones there and the rest in the corresponding code file. This
speeds up building a bit. Then, debug mode will only work when using
this program from the CLI, i.e. not in website mode. And in website mode
don't include thread IDs and filenames in error messages.