Only add image to queue if it was loaded
Previously if an "image" was found in the PDF that wasn't ever
successfully loaded (e.g. the width or height wasn't specified in the
PDF), the Pixels object be blank and still added to be processed, which
obviously was unsuccessful and a waste of time. It errored giving not
enough boxes detected, but didn't specify the filename. Now we just
don't process an image if it isn't loaded.