make install only installs instead of configures
Using cmake you can generate the keys for the website, but with make
install it will only install the files that are the same for all users.
This is to ease packaging, for example on Arch Linux with a PKGBUILD.
The install is the same for everybody but in a freetron.install file the
keys are generated per install so nobody has the same keys.