Keep sorted list of points we've been to in Outline
It is faster to detect if we've been to a point before if we have a
sorted list, so now use std::set in addition to the ordered by
last been to list in a std::vector. This uses more memory, but if the
path grows to be very large (e.g. 100,000 points or so), this will make
a considerable difference.

Note: may not matter much here since we probably don't have that long of
outlines in these black and white images, but it makes a huge difference
when tracing blobs in photos.